bug: 머지 후 빌드 안되는 현상 해결(#28)
