+++ Projects/Feature/Sources/ObservedObject/MainDetailViewObservedObject.swift
@@ -7,20 +7,13 @@
 //
 
 import Foundation
-import SwiftUI
+import Core
 
 class MainDetailViewObservedObject: ObservableObject {
-    
+
     @Published var playerStatistics = PlayerStatistics(totalDistanceCovered: 55,
                                                        totalCalories: 16500,
                                                        totalPlayTime: Date(),
                                                        totalPlay: 11)
-    
-}
 
-struct PlayerStatistics {
-    let totalDistanceCovered: Double
-    let totalCalories: Double
-    let totalPlayTime: Date
-    let totalPlay: Int
 }
